move test folder outside src
