+++ package.json
@@ -22,7 +22,7 @@
   },
   "scripts": {
     "build": "hardhat compile && tsc -p tsconfig.build.json",
-    "check-types": "tsc --noEmit -p src/__tests__/tsconfig.json",
+    "check-types": "tsc --noEmit -p test/tsconfig.json",
     "check-types:incremental": "npm run check-types --incremental",
     "coverage": "c8 yarn test",
     "eslint:check": "eslint . --max-warnings 0 --ext .js,.jsx,.ts,.tsx",
