+++ test/domain-registry.spec.ts
